+++ src/db/migrations/migrations.js
@@ -4,13 +4,15 @@ import journal from './meta/_journal.json';
 import m0000 from './0000_lists.sql';
 import m0001 from './0001_items.sql';
 import m0002 from './0002_shops.sql';
+import m0003 from './0003_item_note_urgent.sql';
 
   export default {
     journal,
     migrations: {
       m0000,
 m0001,
-m0002
+m0002,
+m0003
     }
   }